.archive/backups/2026-02-28_pre-0.8/src/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
aiohttp==3.9.5
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
